+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added support for style `extends`
 - Added support for `verticalAlignment` on table cells
 - Fixed invalid table layout styles for server-side rendering
+- Fixed styling issues around `0`, `undefined` and `null` values
 
 ## v0.3.2 - 2026-01-31